pyproject.toml
[tool.poetry]
name = "awesome-aosp-rust"
version = "0.1.0"
description = ""
authors = ["Pierre Tardy <pierre.tardy@renault.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "awesome_aosp_rust"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
